Moved experiments table logic to custom hooks
