+++ DashAI/front/src/pages/experiments/constants/index.js
@@ -1,0 +1,1 @@
+export * from "./table";
