+++ DashAI/front/src/pages/experiments/hooks/useExperimentsDelete.js
@@ -1,0 +1,22 @@
+import { useState } from "react";
+import { deleteExperiment as deleteExperimentRequest } from "../../../api/experiment";
+
+export default function useExperimentsDelete({ onSuccess, onError }) {
+  const [loading, setLoading] = useState(false);
+
+  const deleteExperiment = async (id) => {
+    setLoading(true);
+    try {
+      await deleteExperimentRequest(id);
+
+      onSuccess && onSuccess();
+    } catch (error) {
+      console.error(error);
+      onError && onError();
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return { deleteExperiment, loading };
+}
